WEB/requirements.txt
torch==2.5.1
torchvision==0.22.1
pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pytorch-grad-cam==1.4.8
gradio==3.40.0
scipy==1.11.1
